+++ src/templates/bootstrap/Size.js
@@ -1,0 +1,5 @@
+import t from 'tcomb-validation';
+
+const Size = t.enums.of('xs sm md lg', 'Size');
+
+export default Size;