+++ src/templates/bootstrap/getError.js
@@ -1,0 +1,9 @@
+import bootstrap from 'uvdom-bootstrap';
+
+export default function getError({hasError, error}) {
+  if (!hasError || !error) { return; }
+  return bootstrap.getErrorBlock({
+    error,
+    hasError
+  });
+}