+++ src/templates/bootstrap/getHelp.js
@@ -1,0 +1,9 @@
+import bootstrap from 'uvdom-bootstrap';
+
+export default function getHelp({help, attrs}) {
+  if (!help) { return; }
+  return bootstrap.getHelpBlock({
+    help,
+    id: attrs.id + '-tip'
+  });
+}