+++ src/templates/bootstrap/getLabel.js
@@ -1,0 +1,15 @@
+import bootstrap from 'uvdom-bootstrap';
+
+export default function getLabel({label, breakpoints, htmlFor, id, align}) {
+  if (!label) { return; }
+
+  const className = breakpoints ? breakpoints.getLabelClassName() : null;
+
+  return bootstrap.getLabel({
+    align,
+    className,
+    htmlFor,
+    id,
+    label
+  });
+}